chore: move v1 contracts under v1 subfolder
