+++ contracts/v1/protocol/generate.go
