+++ contracts/v1/tokens/generate.go
