+++ relayer/client.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 
-	"github.com/linki/0x-go/contracts/protocol"
+	"github.com/linki/0x-go/contracts/v1/protocol"
 	"github.com/linki/0x-go/types"
 	"github.com/linki/0x-go/util"
 )
